pyproject.toml
[project]
name = "bowl-of-scenes"
version = "0.1.0"
description = "GUI application to create batch of images with AI"
authors = [
    {name = "Emmanouil Ragiadakos"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite==0.21.0",
    "mashumaro>=3.17",
    "nicegui>=3.4.0",
    "pytsterrors>=0.4.1",
    "tortoise-orm==0.25.1",
    "yet-another-comfy-client",
]
license = {text = "SSPLv1"}

[tool.uv.sources]
yet-another-comfy-client = { git = "https://github.com/rm4n0s/yet-another-comfy-client" }
